animation処理を切り分け
