+++ src/public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=68bbc46087583b2968f9",
+    "/js/app.js": "/js/app.js?id=a09088b796857a65c85b",
     "/css/app.css": "/css/app.css?id=80e4240a46c3d0b667f5"
 }
