+++ src/resources/js/config/resize.js
@@ -1,6 +1,6 @@
 /**
- * [Windowリサイズに関する処理]
- * 使用したい箇所でmixin登録する。
+ * [Windowリサイズ処理]
+ * 使用コンポーネントでこのファイルをmixin登録する。
  */
 
 export default {
