+++ src/resources/js/config/scroll.js
@@ -1,6 +1,6 @@
 /**
- * [スクロール量取得処理]
- * 使用箇所でmixin登録する。
+ * [スクロール取得処理]
+ * 使用コンポーネントでこのファイルをmixin登録する。
  */
 
 export default {
