+++ src/resources/js/views/History.vue
@@ -89,9 +89,12 @@
 // data
 import Data from '../config/data.json';
 
-// mixin
-import Risize from '../config/resize';
+/**
+ * mixin
+ * resize.js は animation.js に組み込まれているので無駄にインポートしない。
+ */
 import Scroll from '../config/scroll';
+import Animation from '../config/animation';
 
 // import components
 import ContentsTitle from '../components/modules/ContentsTitleComponent';
@@ -101,7 +104,7 @@ import ChampionsCard from '../components/modules/card/ChampionsCardComponent';
 import ScrollTopButton from '../components/modules/button/ScrollTopButtonComponent';
 
 export default {
-  mixins: [Risize, Scroll],
+  mixins: [Scroll, Animation],
 
   components: {
     ContentsTitle,
@@ -244,18 +247,6 @@ export default {
       // 沿革ボックスのmargin-bottom を数値で取得
       this.archiveMarginBottom = parseInt(window.getComputedStyle(this.$refs.taisho).marginBottom);
     },
-
-    fade(evt, el) {
-      let top = el.getBoundingClientRect().top;
-      let trigger = this.windowHeight / 2;
-
-      if (top < trigger) {
-        el.setAttribute(
-          'style',
-          'opacity: 1;'
-        )
-      }
-    }
   },
 
   watch: {
