+++ src/resources/js/views/Home.vue
@@ -53,6 +53,9 @@
 // data
 import Data from '../config/data.json';
 
+// mixin
+import Animation from '../config/animation';
+
 // component import
 import MainVisual from '../components/contents/MainVisualComponent';
 import News from '../components/contents/NewsComponent';
@@ -68,6 +71,8 @@ export default {
     LinkButton,
     ScrollTopButton,
   },
+
+  mixins: [Animation],
 
   data() {
     return {
@@ -100,13 +105,6 @@ export default {
         if (el.title === element.key) el.title = element.value;
       });
     });
-  },
-
-  mounted() {
-    this.$refs.about.forEach((element) => {
-      let top = element.getBoundingClientRect().top;
-      console.log(element);
-    })
   },
 
   methods: {
